requirements.txt
streamlit
pandas
langchain
langchain-openai
langchain-core
python-dotenv
gspread
oauth2client
plotly